requirements.txt
fastapi~=0.70.0
uvicorn~=0.15.0
pandas~=1.3.3
python-dotenv~=0.19.0
pytest~=7.4.3